Environment start empty, so that bootstraping env is inlined directly.
Fixes by accident the problem of mismatch between getNodes and type
signatures. A mismatch can exist between the type specified in the
signature and the declared type of the ref. To mitigate that issue we
will propagate type signatures for rec/args up to depth 1, and all
localvariable except receivers.