Handle exceptions a tad bit better
Instead of yielding an incomplete CFG that will not allow for proper
retval mapping, it now generates a CFG where exit is potentially
unreachable. If it is, it's flat effect is de-facto Bottom and all is
good. We can disable the warning in case all targets are Bottom'ish,
i.e. Nil.head.